Weather, MainActivity and SettingsSetter update!
New asyncTask added to get the weather.
It's disabled by default but can be enable from SettingsSetter.
When clicked on inside SettingsSetter it will open a CustomAlertDialog.
In the alert dialog the user can pick between capitals around the world.
The weather going to be about the capital that selected.
Weather auto update added but not yet to be tested.
alert_dialog_find_capital.xml added.

MyAsyncClass renamed to UploadInfoTask.

Some more changes related to the new weather option.