Plan, User, DailyMenu and HomeFragment update!
Plan got some new features, including Plan history called previousPlans inside of the User class.
Plan got two new properties "fromDate" and "untilDate" to support this action.
Plan got a new constructor of Plan(Datasnapshot datasnapshot, int mode).

User now have the property "previousPlans".
It will need to be updated if not yet to be when internet connection available.

Home screen now include a button that allow the user to pick the date of menu he want to see.
This pick wont be changed until he exit the app and reenter it.

StatisticsFragment has added but not yet been implemented.
A try in making it include several pie charts is on his way.

Song got a little change to prevent crashing in certain cases.
DailyMenu got a little change to prevent crashing in certain cases.