Final update!
The app got a lot of visualization changes to improve user-experiment...
A lot of new tests has added to insure the app running crash-free!
A new video for the UserInfoScreen has added.

Fixed a bug that cause crash after registered to the app and attempt to add food.
Fixed a bug that cause crashes some times while using the app.
Fixed a bug with the Plan.isTheSamePlan function.
Fixed some crashes that caused by int limits.

Some more improvements to insure user-friendly experiment.